+++ inst/shiny-examples/BioMonTools/ui.R
@@ -21,9 +21,8 @@
 
 dashboardPage(
   header = dashboardHeader(title = "BioMonTools")
-  #, sidebar = dashboardSidebar(sb_main("leftsidebarmenu"))
-  , sidebar = dashboardSidebar(sb_main("leftsidebarmenu"))
-  , body = dashboardBody(db_main("dbBody"))
+  , sidebar = dashboardSidebar(main_db_sb("leftsidebarmenu"))
+  , body = dashboardBody(main_db_body("dbBody"))
 ) ## dashboardPage ~ END
 
 # https://rstudio.github.io/shinydashboard/get_started.html